Fixes to the back-end
