+++ lib/package.json
@@ -2,7 +2,7 @@
   "name": "lib",
   "version": "1.0.0",
   "description": "",
-  "main": "upgradePath.js",
+  "main": "app.js",
   "dependencies": {
     "@octokit/rest": "^16.35.0",
     "body-parser": "^1.19.0",
@@ -12,11 +12,12 @@
     "express": "^4.17.1",
     "get-repo-package-json": "^2.0.0",
     "npm-check": "^5.9.0",
-    "parse-yarn-lock": "^999.999.2"
+    "parse-yarn-lock": "0.28.4"
   },
   "devDependencies": {},
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node app.js"
   },
   "author": "Andy Tsang",
   "license": "ISC"
