+++ lib/routes/upgradePath.js
@@ -4,15 +4,9 @@
 'use strict'
 
 const express = require('express');
-const port = 8000;
-const app = express();
+const router = express.Router();
 const npmCheck = require('npm-check');
 const https = require('https');
-const bodyParser = require('body-parser')
-
-app.use(bodyParser.urlencoded({
-    extended: true
-}));
 
 const jsonLocation = './input/';
 const fs = require("fs");
@@ -128,7 +122,7 @@ function retrieveSizes(libraries) {
 /*
  * Retrieve the information of all package.json libraries
  */
-app.post('/upgrade', (req, res) => {
+router.post('/upgrade', (req, res) => {
     res.status = 200;
     // console.log(req.query);
     retrieveJson(req.query.repo).then(() => {
@@ -140,7 +134,7 @@ app.post('/upgrade', (req, res) => {
 /*
  * Retrieve the size information of a list of libraries 
  */
-app.post('/size', (req, res) => {
+router.post('/size', (req, res) => {
     res.status = 200;
     // console.log(req.query.dep);
     retrieveSizes(JSON.parse(req.query.dep)).then(() => {
@@ -152,7 +146,7 @@ app.post('/size', (req, res) => {
 /*
  * Retrieve the git commit information of the package.json
  */
-app.post('/commit_history', (req, res) => {
+router.post('/commit_history', (req, res) => {
     res.status = 200;
     sendPostGithub(req.query.repoUser, req.query.repoName).then((result) => {
         res.send(result);
@@ -196,10 +190,6 @@ function sendPostGithub(repoUser, repoName) {
     });
 }
 
-app.listen(port, () => {
-    console.log('Example app listening on port 8000!')
-});
 
-
-
+module.exports = router;
